added the check for Leo saturday monitor role

+++ bot.js
@@ -4,6 +4,7 @@ const Discord = require('discord.js');
 const client = new Discord.Client();
 const fetch = require('node-fetch');
 
+const prefix = '!'
 client.login(process.env.BOT_TOKEN);
 
 client.on('ready', () =>{
@@ -13,8 +14,8 @@ client.on('ready', () =>{
 
 async function getGif(tema)  {
   
-  let cacafuuGIF = `https://g.tenor.com/v1/search?q=${tema}&key=${process.env.TENORKEY}`
-  let response = await fetch(cacafuuGIF);
+  let gifRequest = `https://g.tenor.com/v1/search?q=${tema}&key=${process.env.TENORKEY}`
+  let response = await fetch(gifRequest);
   let json = await response.json();
 
   const randomIndex = Math.floor(Math.random() * json.results.length)
@@ -30,15 +31,15 @@ client.on('voiceStateUpdate', async function (oldMember, newMember)  {
   const textChannel = client.channels.cache.get('322106045486858240');          //Text channel ID
 
 
-  if(newUserChannel === '419612592542318593' && oldUserChannel === null) {      //Voice channel ID & alguem entrou no voice
-    textChannel.send(`Tamo saindo galera`,{tts: true}).then((message) => {
-      setTimeout(() => {          //Apaga mensagem após 3 Minutos
-        message.delete();
-      }, 1000 * 180)
-    });
+  if((newUserChannel === '795049118304436234' || newUserChannel === '862731867295711232' ||  newUserChannel === '847310147478355969') && oldUserChannel === null) {      //Voice channel ID & alguem entrou no voice
+    // textChannel.send(`Tamo saindo galera`).then((message) => {
+    //   setTimeout(() => {          //Apaga mensagem após 3 Minutos
+    //     message.delete();
+    //   }, 1000 * 180)
+    // });
     if(newMember.id ==='256839511781277696'){
       
-      textChannel.send(`É o homem macaco`,{tts: true}).then((message) => {
+      textChannel.send(`É o homem macaco`).then((message) => {
         setTimeout(() => {        //Apaga mensagem após 3 Minutos
           message.delete();
         }, 1000 * 180)
@@ -50,11 +51,19 @@ client.on('voiceStateUpdate', async function (oldMember, newMember)  {
       });
 
     } else if (newMember.id ==='260557019428290570') {        //LeoDancer
-      textChannel.send(`Chegou otaku`,{tts: true}).then((message) => {
-        setTimeout(() => {          //Apaga mensagem após 3 Minutos
-          message.delete();
-        }, 1000 * 180)
-      });
+      let today = new Date();
+      if(today.getDay() == 6 && (today.getHours() > 4 && today.getHours() < 11)){
+
+        textChannel.send(`Nosso segurança chegou no horário, bateu o ponto às ${today.toLocaleTimeString()} e está no horário`, {files: ["https://cdn.discordapp.com/attachments/322106045486858240/865922496901808158/20210717_084633.jpg"]});
+
+      } else {
+    
+        textChannel.send(`Chegou otaku`).then((message) => {
+          setTimeout(() => {          //Apaga mensagem após 3 Minutos
+            message.delete();
+          }, 1000 * 180)
+        });
+      }
 
       textChannel.send(await getGif("weeb")).then((message) => {
       setTimeout(() => {            //Apaga mensagem após 3 Minutos
@@ -81,7 +90,7 @@ client.on('voiceStateUpdate', async function (oldMember, newMember)  {
     }
 
   } else if (oldUserChannel === '419612592542318593' && newUserChannel !== '419612592542318593') { // Alguém saiu do voice
-    textChannel.send(`Agora a gente ganha`,{tts: true}).then((message) => {
+    textChannel.send(`Agora a gente ganha`).then((message) => {
       setTimeout(() => {        //Apaga mensagem após 3 Minutos
         message.delete();
       }, 1000 * 180)
@@ -89,6 +98,17 @@ client.on('voiceStateUpdate', async function (oldMember, newMember)  {
   }
 })
 
+
+client.on('message', message =>{
+  if(!message.content.startsWith(prefix) || message.author.bot) return
+
+  const args = message.content.slice(prefix.length).trim().split(/ +/);
+	const command = args.shift().toLowerCase();
+
+	if (command === 'sexo') {
+		message.channel.send('O sexo é um acidente: o que dele recebemos é momentâneo e casual; visamos a algo mais secreto e misterioso do qual o sexo é apenas um sinal, um símbolo.Pong.');
+	}
+})
 //Test
 // client.on('voiceStateUpdate',async function (oldMember, newMember)  {
 //   const newUserChannel = newMember.channelID;
@@ -102,8 +122,16 @@ client.on('voiceStateUpdate', async function (oldMember, newMember)  {
 //   if(newUserChannel === '693310952476639242' && oldUserChannel === null) {
 //     textChannel.send(`Tamo saindo galera`,{tts: true});
 //     if(newMember.id ==='257938615907123201'){//Para Cacafuu
-//       textChannel.send(`É o homem macaco`,{tts: true})
-//       textChannel.send(await getGif("macaco"))
+//       let today = new Date();
+//       if(today.getDay() == 6 && (today.getHours() > 0 && today.getHours() < 11)){
+
+//         textChannel.send(`Nosso segurança chegou no horário, bateu o ponto às ${today.toLocaleTimeString()} e está no horário`, {files: ["https://cdn.discordapp.com/attachments/322106045486858240/865922496901808158/20210717_084633.jpg"]}).then((message) => {
+//           setTimeout(() => {          //Apaga mensagem após 3 Minutos
+//             message.delete();
+//           }, 1000 * 180)
+//         });
+
+//       }
       
 //       // console.log(await getGif("macaco"))
 //     }
@@ -111,3 +139,13 @@ client.on('voiceStateUpdate', async function (oldMember, newMember)  {
 //     textChannel.send(`Agora a gente ganha`,{tts: true})
 //   } 
 // })
+// let today = new Date();
+//       if(today.getDay() == 6 && (today.getHours() > 4 && today.getHours() < 11)){
+
+//         textChannel.send(`Nosso segurança chegou no horário, bateu o ponto às ${today.toLocaleDateString} e está no horário`, {files: ["https://cdn.discordapp.com/attachments/322106045486858240/865922496901808158/20210717_084633.jpg"]}).then((message) => {
+//           setTimeout(() => {          //Apaga mensagem após 3 Minutos
+//             message.delete();
+//           }, 1000 * 180)
+//         });
+
+//       }
